requirements.txt
langchain==0.2.16
langchain-openai==0.1.23
python-dotenv